front: extract helper function from useOccurrences hook
